avances en programacion de calculos que tienen que ver con los combobox, pero hay problemas con el calculo de valor de cuotas. Investigando key events y listeners

+++ src/Convenios/Convenio.java
@@ -21,7 +21,7 @@ public class Convenio {
         this.nombre = nombre;
     }
 
-    public float getPorcentaje() {
+    public int getPorcentaje() {
         return porcentaje;
     }
